(SDK-191) Allow validators and targets as arguments rather than options
This commit updates the `validate` subcommand to take validations
and targets as arguments rather than options.

If a user wishes to specify which validations to run, they must
supply a comma separated list of validators as the first argument,
before any targets are specified.

* If a comma separated list if provided as the first argument, we
assume it is a list of validators.
* If a single item is provided as the first argument, we check if
it is a known validator. If so, we treat it as such. Otherwise, we
treat it as a target.
* All subsequent arguments are treated as targets.