(SDK-147) Implement 'test unit' subcommand
This commit implements basic support for the `test unit` subcommand
which invokes the puppetlabs_spec_helper provided `rake spec` command
inside the context of the target module's bundled gem environment.
Output from the underlying RSpec command is parsed and then returned to
the user in the requested format and to the requested destination.