(MAINT) Update to official master of github-changelog-generator
Since skywinder/github-changelog-generator#550 was merged, we can switch to the upstream repo.

Once a new version is released we can drop the git reference altogether.

This also includes a minor config change to make the rake task work, and the current output to show the minimal changes to the CHANGELOG file's format, and no content changes in the historical entries.