(PDK-420) Ensure Puppet and Puppet::Util modules are defined
Puppet 5 changed the format of the module definition in Puppet::Util::Windows
from the expanded to condensed format. The downside of this change is that it
means that the requires are now order dependent, as the condensed format does
not define the parent objects if not already defined.

This change abstracts our require of `puppet/util/windows`, allowing us to
define the `Puppet` and `Puppet::Util` modules before requiring
`puppet/util/windows`, fixing the exception caused when `pdk` and `puppet`
exist in the same gemfile.