(PDK-450) remove stdlib dependency
Having the stdlib dependency in there, but not in the rspec-puppet fixtures
leads to painful expectations: stdlib functions are available, but using
them in the code leads to test failures. Adding it to the fixtures would
cause the default PDK workflow to require a online connection, so this
instead removes the unforced dependency.