(MAINT) Run package test commands in a login shell
Beaker defaults to running commands in a non-interactive /bin/sh session
which means that /etc/profile and other shell setup files are not
sourced. This commit changes our helper method to still run 'pdk'
without an explicit path but do so inside a 'login' session of /bin/bash
so that any profile based PATH changes are in place when the command
runs.