(PDK-719) Check for module directory layout when converting
Falls back to checking for `manifests/`, `lib/`, `tasks/`, `functions/`, and
`types/` if the module does not contain a `metadata.json`. This logic is not
enabled by default as we should only need to do this during convert (after
that, a `metadata.json` file should be present).

`templates/`, `files/`, `spec/` are not checked for, as it's unlikely they'll
exist without one of the other directories being present.