(MAINT) Update to released version of GCG
This version contains the label filtering that we need, and adds
more info to the changelog. Since we're maintaining separate release
notes on puppet.com/docs anyways, this seems to be a benefit, as it
gives multiple levels of details. Added a link to the main release notes
to connect them.