(PDK-736) Improve handling of old template-url and template-ref
Previously, if the user had the old puppetlabs/pdk-module-template URL
saved in their answers.json, PDK would still try to use the old repo but
attempt to reset to the `1.3.0` tag which doesn't exist in the old repo.
This commit starts treating a value matching the old repo URL the same
as if there is no template URL specified so that it uses the new
template URL instead.

Furthermore, if the user had a real custom template repo, we were still
attempting to reset to the `1.3.0` tag which probably won't exist in
their custom repo. After this commit, pdk will only try to reset to the
pinned tag when using the official template repo (or packaged repo).