Add --template-ref argument for upstream template repo tags
I rolled the url/refs together internally into a single "uri" that
should be considered the template target and broken apart with the URI &
decode methods when needed separately.

- The decision code for obtaining the uri/ref has been centralized into
the `PDK::Util` functions as discussed on the PR comments on github.

- The template-ref value in the metadata is for pdk update reference,
not for template targetting.

- The default ref is to be used when no fragment is available in the
URI.

- The `PDK::Util.template_uri()` takes options hash for the opts.fetch
in case users pass --arguments, but is a class method. I don't really
like this, but it's probably needed to fulfill the requirements
discussed in the PR.

- The `file:///...` urls are treated as remote paths, but absolute paths
such as `c:/...` or `/...` are treated as local paths and use directly.
This behavior was pre-existing.

- The work tree for the templates directory is only reset newly-cloned
directorys, and should not be touched in the above-mentioned local-path
directories.

- Handling file paths with #'s in the name is overly complex due to the
way URI encoding and URI fragments work. I just error in the case of #'s

- The Addressable::URI library is the most-compatible URI handler, so I
used that.

- SCP-like urls of user@host:/path are not URI RFC-compatible, so I
helpfully parse them into URIs and print a warning. (The ambiguous case
is `user@host:<0-65535>/path` in which I interpret this as a URI instead
of an SCP url)

- I removed the --template-url --template-ref arguments from most of the
`pdk new *` generators as they are not used to specify or change
templates and actually create edge cases by being used to do so. The
arguments should only be specified during a pdk convert or a `pdk new
module`

- Most of the spec fixes only have to do with shuffling a few stubs
around and lining up string/addressable::uri expectations.

- One awkward thing I noticed is that the spec_helper.rb on line 33
loads the answers by requiring pdk/cli and it happens before any stubs
are placed so polluted answers.json on my machine is picked up instead
of stubs. This should be documented probably as I lost a few hours to
it.

- Windows paths most start with either a `file:///<drive>:/...` scheme
or `/<drive>:/` leading-slash to be URI-compatible. I represent paths
that lack the `file://` scheme as `/<drive>:/` internally and strip the
leading / when used for commands or serialization in metadata.json.