(PDK-802) Work around OpenSSL multi-threading errors when needed
See https://bugs.ruby-lang.org/issues/11033 for background.

TL;DR is that there was a bug in the Ruby 'openssl' gem that caused a
race condition on Windows when loading certs.

I think this is the cause of "cert already in hash table" transient
failures we see in Appveyor as well as some of the issues @glennsarti
was encountering in https://tickets.puppetlabs.com/browse/PDK-802

Ruby issue linked above was backported to Ruby 2.3 branch and released
in 2.3.5, however Appveyor is still on 2.3.3 for Ruby23 target.

Inside PDK packages it should only be an issue on the Ruby 2.1.9
bundler invocations.