(FIXUP) Make convert acceptance tests use same template as generate
The `pdk convert` acceptance tests were not specifying a --template-url
which caused them to fall back to reading whatever template-url was
specified in the users answers file, rather than the same template-url
that was used to generate the test module.

Also the tests were expecting there to be a .project file to manage but
that was recently removed from the templates.