(maint) do not test branches other than master
When pushing a PR source branch directly to the puppetlabs repo,
e.g. when using the Github UI to make a quick fix, travis and
appveyor would run the tests twice. This prevents that by restricting
builds to `master` only. This preserves post-merge checking of the
main branch to make sure that we do not miss merge problems.