(PDK-949) Add merge options to the .sync.yml
Prior to this commit the config_defaults.yml and .sync.yml were deep
merged without any options. This commit adds in the ability to pass in a
`merge_options` hash in the `.sync.yml` to specify options for merging
the hashes. The main use case for this is to add a knockout prefix for
removing some of the `config_defaults.yml` in the template.