(PDK-949) Add a default knockout_prefix for options
Prior to this commit the config_defaults.yml and .sync.yml were deep
merged without any options. This commit adds in the ability to use a
knockout_prefix of '---' for any config_defaults.yml options.
The main use case for this is to remove items from the
`config_defaults.yml` in the template.