(PDK-1061) Ensure rake binstub when building module
Before building a module package, we first call
`PDK::Test::Unit.tear_down` to make sure that there are no artifacts
left over from possible test runs. This method depends on the `rake`
binstub being present which is ensured when running `pdk test unit`,
however when the user is running `pdk build` on a clean clone of
a module this binstub is not present and the build fails.