(PDK-1048) Improve docs for `pdk test unit --verbose`
Improve the help output for `pdk test unit`'s `--verbose` flag to make
it clearer that it only affects the behaviour when used in conjunction
with `--list`. Additionally, I've added an info level log message if
`--verbose` is used without `--list` to explain that it does nothing in
this case.