(PDK-1045) Send validation targets as relative file paths
Update the validator to parse targets and return results as file paths
relative to the module root. This is to cut down on the character length
of the command string, since the full path of the module directory is
unnecessary and long module_root paths may end up hitting the maximum
character limit on Windows.