(PDK-1088) Remove unnecessary file enumeration loop during PDK build
Prior to this change, every call to PDK::Module::Build#ignored_files
resulted in a Find.find enumeration of the entire module tree to test
whether or not the target_dir should be added to the ignored_files
PathSpec.

The fundamental issue was that this enumeration happened on
every call when it should have been cached after the first call, but I
also refactored the check itself to use File.realdirpath and simple
string comparisons.