(PDK-400) Output the rspec run wall time in test unit summary
Currently `pdk test unit` outputs the time rspec took to evaluate the
tests in the summary (which is always significantly shorter than the
time taken to execute the command) while `pdk test unit --parallel`
outputs the wall time for parallel_spec. This leads to a confusing
situation where running tests in parallel seems to be slower than
running the tests sequentially.

This PR changes the behaviour so that the wall time is always printed
(which is, in my humble opinion, the more useful metric in most cases).