(PDK-1093) Replace null values and empty data structures in metadata when converting
Currently when we're updating a module's `metadata.json` file during
convert, we only replace values if they are completely absent from the
file. Unfortunately, this means if - for example - the existing metadata
defines `dependencies` as `null` or an empty array then it will not be
updated when the module is converted and it will subsequently fail to
validate cleanly.

This PR changes the metadata merging logic so that values defined as
`null`, empty array, or empty hash are considered to be undefined and
will be replaced with the default metadata value.