(PDK-914) Adjust default_template_url validation to accept local dirs
Currently, if the template-url specified in the answer file points to
a local directory any calls to `PDK::Util.default_template_url` will
falsely report that the template is invalid even when it is a valid
template. This happens because `PDK::Util.default_template_url` only
expects the answer file to contain a template url pointing at a git
repository.