(PDK-1193) Saves packaged template-url in metadata as a keyword
Currently, for packaged installations the default templates template-url
is saved in the metadata.json as the file path. This breaks when a
module that is developed with PDK on Windows is moved to a Linux installation
of PDK, since the Windows path doesn't exist. With this fix, the default
packaged template path will be determined by PDK using the keyword
'pdk-default' in that same metadata.json field and the approriate
platform specific template location will be used.