Update logic for updating templates between tags
If the module being updated is using the default template *and* the
template ref is a tag *and* the PDK is being run from a package install,
then update the module to use the latest tag of the template. In all
other cases, get the ref from the `template_url` metadata value (and if
there is no ref in the `template_url` metadata value, use the default
template ref which is the latest tag for package installs or `master`
otherwise).