(PDK-1308) Ensure PDK-written non-templated files have trailing newline
There are some files that PDK writes that are not derived from templates
in the pdk-templates repo. The most notable example of this is the
metadata.json file which is constructed as a Ruby hash and then
serialized to JSON and written to a file. When PDK writes these files it
was not including a trailing newline and files having trailing newlines
is a best practice.

This change updates the PDK::Util::Filesystem.write_file method to
ensure that it always writes files with a trailing newline.

Some of the unit tests around writing templated files also needed to be
updated because of the way the tests were implemented did not simulate
template files that already contained trailing newlines.