(MAINT) Temporarily disable using puppetlabs-motd for package testing
puppetlabs-motd now includes puppet_litmus in it's dependencies, and
puppet_litmus as currently released depends on 'bolt' and 'pdk' gems. It
probably shouldn't depend on those things as a library and it looks like
that is the direction that the puppet_litmus team is moving, but until
they have released a gem that removes those deps, we should stop using
litmus modules for package testing.