(maint) Pin cri to <= 2.15.6
cri 2.15.7 changed the behaviour of the options hash so that unspecified
flags are defaulted to an explicit `false` rather than being unset. From
the look of https://github.com/ddfreyne/cri/issues/94 though this
behaviour may be changing in the next release so lets pin cri to
a maximum of 2.15.6 for now until the behaviour stabilises and then we
can make the necessary changes.

/cc #674