(PDK-1365) Use dynamic ruby detection for default ruby instance
Previously the default version of ruby selected in packaged PDK instances was
hardcoded to 2.4.5 which caused issues when updating ruby runtimes and did not
reflect that the default should really by ruby 2.5 and Puppet 6.  This commit
modifies the default ruby detection to instead use the ruby environment for the
latest Puppet Gem in the PDK package install.  Non-pkged installs are not
affected.

This commit also memoizes the default value to slightly speed up default ruby
selection.