(maint) Clear Gemfile overrides before pdk update test
The two modules we test `pdk update` with in the package tests pull in
additional modules which then pull in unsupported gems with compiled
extensions, causing the tests to fail.

By removing these gems from the `Gemfile` before running the tests, we can
get the tests working again while retaining the original purpose of the
tests (testing that packages that depend on beaker can be updated and
run without needing a compilation environment).