(PDK-1432) Autogenerate PowerShell modules from code
Previously the PowerShell module code lived in the pdk-vanagon repository
however during release this required manual steps.  This commit moves the
PowerShell module files to this repository and allows the use of ERB to update
them, for example, version and copyright year.

This commit also adds a beta PowerShell module which provides a much richer
and native PowerShell experience, by autogenerating from the CRI information.