(PDK-1434) Gracefully handle unparsable bolt analytics config
This issue was found because the classroom Windows instances created the
`~/puppetlabs/bolt/analytics.yaml` manually with UTF-16LE encoding which
Psych on Windows is unable to parse, causing
`PDK::Config.bolt_analytics_config` to raise `PDK::Config::LoadError`.

The classroom bootstrap scripts have been fixed to generate the file as
UTF-8 now, but we should still gracefully handle this case and have PDK
behave as if the file doesn't exist and use the default values.