(PDK-1107) Add pdk config get CLI command
Previously there was no way to display what the current configuration the PDK
was using.  This commit adds a `pdk config` command, with a single `get`
sub-command.

This commit also adds the ability to filter the settings either by a specific
name or setting treename.

This commit also adds tests for the new resolve method in the Namespace object.