(GH-722) Do not emit nil targets for validators against a directory
Previously if the user passed a valid directory to a validator which did not
have a pattern glob which included that directory it emitted nil targets which
then broke the validator when it tried to actually use nil.

This commit:
* Changes the directory target list from a map to a select-map to only emit
valid directory targets, instead of emitting nils.
* Removes the debug message as the actual message was incorrect. Also the file
based codepath doesn't emit debug messages, so this brings the codepath into
alignment with that.
* Adds a unit test for the scenario of passing file and directory targets into
a validator with a pattern which doesn't match any of the targets.