(PDK-1112) Add config namespaces which use JSON Schema
Previously there was no formal schema, or a way to validate a schema, for
configuration settings. This commit adds two new namespace types, YAML and
JSON files with JSON schema validation.

These namespaces will also take a JSON Schema file and use that to provide
defaults and validation.  Any settings NOT in the schema will not be visible in
the namespace, but will not be lost when saving the namespace to disk; That is
the namespace will only modify settings that are in the schema. This also
includes tests for any schema files, to ensure that they are indeed valid.

This commit adds a schema for the PDK analytics configuration.

This commit introduces a new setting type, JSONSchemaSetting which uses the
schema to validate and get default values for a setting from the schema.