(PDK-1525) Use less restrictive minitar version
Previously the PDK used the minitar restriction similar to that of the Puppet
Windows build (~> 0.6.1).  However, from Puppet 6.10.1 this version pin
changed to `~> 0.9`. This means that using PDK as a gem on Windows, is causing
bundler to choose old versions of gems (e.g. doesn't include Litmus).

This commit changes the minitar pin to still be compatible with older Puppet
gems and still allow later Puppet too.