(MAINT) Fix new provider error message
Prior to this commit, if a user creates a module with
the PDK, then tries to create a provider, they get an
error - but the error leads to an anchor on the docs
for the resource_api readme which does not exist.

This commit corrects the URL of the error message,
pointing the user to the actual instructions they need
to follow to move forward.