(GH-768) Fix in_module_root? gives false positives
Previously the module root detection used the bare `lib` directory to detect if
a user was in a module root, without a metadata.json.  However this caused
false positives as `lib` is a very common directory. For example when in a ruby
gem project. This commit modifies the detect to use more puppet specific paths:
* lib/puppet
* lib/facter
* lib/puppet_x

This commit also adds tests for these folder directories, and uses the PDK
Filesystem abstraction classes.