(GH-768) Mock module root in tests
Previously the module root detection was changed to not detect the bare `lib`
directory however this caused many of the unit tests to fail as it was detecting
the PDK gem as a valid module (which it is not).  This commit adds an empty
module root fixture and mocks the various tests use this as the module root.