(GH-768) Remove redundant tests
Previously the `ensure_in_module!` tests tested for the various module
directories (lib, tasks and so on).  However this is really just duplicating
the tests for `in_module_root?` which is not required.  This commit removes
these tests and just mocks the response from `in_module_root?`.  This results
in no loss of test coverage.