(PDK-1519) Print deprecation notice on Ruby < 2.4
Happens at the start of the global CLI runner block so that it applies
to all PDK commands.

This doesn't inform users who are purely using PDK as a library though,
but I think that they're better served by documenting the upcoming
deprecation as well as following semver and performing a major version
bump when dropping support for older Rubies.