Update json-schema dependency
Version 2.8.0 of the json-schema gem assumed the WIP draft 6 URI would
be "http://json-schema.org/draft-06/schema#".  It happens that it is in
fact "http://json-schema.org/draft/schema#".  While this issue was fixe
in ruby-json-schema/json-schema#376 and json-schema 2.8.1 was released
with this fix, it does not fix our problem.

But since we do not depend on draft-06 features, downgrade to draft-04
(there is no draft-05).