(PDK-1364) Update git repo detection and add caching layer
Quering if a directory is a git repository is an expensive call, particularly
on Windows. This commit adds a caching layer to the repo? and work_tree? methods
that caches the results for a short amount of time (GIT_QUERY_CACHE_TTL).

These methods are commonly called multiple times when doing PDK Update and PDK
Convert operations (at least twice).

This commit also updates the repo? method to be consistent with its name.
Previously it would return false for local directories that have a working tree.
With this change it will detect if a directory is any kind of git repository.