(PDK-1169) Add VMWare fallback to PDK::Util::Filesystem.mv
`File.rename` (used inside of `FileUtils.mv`) raises `Errno::ENOENT`
when used inside of a VMWare shared folder on Windows. This adds
a fallback to `PDK::Util::Filesystem.mv` that first checks if the source
file exists (in case the error is legitimate) and then "moves" the file
using copy and remove operations.

As this is quite the corner case, there are no tests added but I've
tested it on a local VM and it resolves the issue.