(PDK-1364) Rename TemplateDir.new to .with
Previosuly TemplateDir was a class, but is now a module. While it is still
possible to use a method called `new` this is confusing as it looks like a
normal object instantiation call.  This commit updates the method name to .with
which is a much better description of what the method does.