(GH-808) Safely read the Changelog file
Previously the changelog file was hardcoded to CHANGELOG.MD however this is not
the default filename.  This commit updates the changelog_file method to search
for different file extensions for the Changelog.  Typically this would only
affect case-sensitive file systems.

This commit also updates the changelog_content to return an empty string if the
changelog file could not be found.  This helps modules which are being released
for the first time, that may not have a changelog file yet.