(PDK-1556) Use the module root during convert and update
Previously the module convert and update operations always assumed the user was
in the module root.  This caused many files to be erronously added to modules if
the operation was started deep within the module.

This commit changes the update and convert classes to accept a module_dir
parameter which explicitly tells the converter/updater where the module root is
so that it can make better decisions on module information (e.g. metadata or
spec files)

This commit also updates the tests for the new required parameter.