(PDK-1588) Increase granularity of `pdk bundle` analytics
Up until now, when a user ran a `pdk bundle` command we would only
report a `bundle` screen view to our analytics. This resulted in us
finding out that a lot of users use `pdk bundle` but provided no insight
into why.

When a user runs `pdk bundle`, we will now capture the bundler
subcommand (but not any of the other arguments). The only exception to
this is if the user runs `pdk bundle exec` in which case we will also
capture the name of the command that is being executed (but again, none
of the arguments).

e.g.

 * `pdk bundle show some_gem` will now be reported as a `bundle_show`
    screen view.
 * `pdk bundle exec rspec path/to/some/ruby_spec.rb` will now be
    reported as a `bundle_exec_rspec` screen view.