(PDK-1592) Add ValidatorGroup abstract class
This commit also adds an abstract class for running a group of Validators
(ValidatorGroup).  This abstract class expects a `validators` method to be
overridden which will list an array of Classes.  The classes could be either a
Validator or even another ValidatorGroup, to create a tree of Validators.

This commit also adds tests for the new abstract class.