(PDK-1592) Add InvokableValidator abstract class
This commit adds an abstract class which is a Validator that can invoked against
a list of targets.  This abstract class is responsible for parsing target lists
and setting up the UI for the Validator.

Sub-classes will then implement the actual validation (via the `invoke` method).

This commit also adds tests for the new abstract class.