(PDK-1592) Add internal ruby and external command abstract validators
This commit adds two abstract classes ExternalCommandValidator and
InternalRubyValidator.  These two classes provide all of the work to validate
either; By running an external command in the bundle context or by running a
ruby block internal to the PDK Ruby runtime respectively.

For example the ExternalCommandValidator would be used to run rubocop and the
InternalRubyValidator would be used to parse a JSON file using the JSON ruby
class. Sub-classes will then implement the actual validation.

This commit also adds tests for the new abstract class.