(PDK-1592) Add Tasks validators
This commit adds the tasks validators using the new abstract validation
classes. This code is based mostly on the existing tasks validators.

Unfortunately as the existing validator classes are defined as 'class Tasks'
instead of 'module Tasks', they had to be removed as well.

This commit also adds tests for the new abstract classes.