(PDK-1592) Add YAML validators
This commit adds the yaml validators using the new abstract validation
classes. This code is based mostly on the existing yaml validators.

Unfortunately as the existing validator classes are defined as 'class YAML'
instead of 'module YAML', they had to be removed as well.

This commit also adds tests for the new abstract classes.