(PDK-1590) Use short-name for PDK Package base directory on Windows
Currently the Package Base Directory detection uses the full path to the PDK
installation on Windows, however Bundler is very sensitive to spaces in the
directory name [1] [2].  This commit modifies the pdk_package_basedir method
to convert the path to its short name using the Windows GetShortPathNameW API
call.

Without this commit different Bundler versions on different Ruby versions may
fail to run `pdk bundle ...` style calls.

Note that the short path detection is not fool-proof.  For example if 8.3
filename creation is disabled there is no shortname to find. In that case it
just returns the long name.

This code path is not explictly tested however it is implicitly tested in the
package-testing tests as the PDK package is installed into "C:\Program Files\.."
by default.

Note that paths WITHIN the pdk are currently ok and don't require the API call
as they don't have spaces in them.  It's just the path TO the PDK that could
contain spaces.

[1] https://github.com/rubygems/rubygems/issues/117
[2] https://stackoverflow.com/questions/16898286/error-invalid-switch-in-rubyopt-f-runtimeerror-is-shown-while-install-gems