(PDK-1590) Use short path names when calling Bundler ruby bin
Previously the full path was used to call the Bundler bin. Unforunately as in
commit 10736a4e4 Bundler is sensitive to spaces in path names.  This commit
modifies the bundler bin path generation to use short path name.

Note that this does not have unit tests as it can't really be tested. But it is
tested as part of the packaging tests.

Note that an additional Windows detection method (PDK::Util.on_windows?) is
introduced.  This method is designed specifically to _NOT_ be mocked and any
code using this method will probably raise errors on non-Windows platforms e.g.
calling Windows API via FFI.  Gem.win_platform? is commonly used in the PDK
codebase and is also commonly mocked.