(maint) Update License Finder
Previously the Travis tests for license_finder were breaking in master due to
Travis configuration. Also the gem was pinned to a very old version.  This
commit updates the Travis configuration so that it correctly sets the bundler
args (can't use an empty string anymore) and updates the license_finder gem
to the latest that's compatible with Ruby 2.1.9.