(maint) Fix namespace file loading
Previously the load_data method took a single string as the name of the file to
load, however the actual file loaded was the file attribute of the class. This
commit modifies the file loading to actually load the filename passed.

This hasn't been an issue as yet, as both the filename and file parameters have
been the same.