(maint) Ensure that setting values are safe to modify
Previously it was possible to get a current PDK configuration setting and then
accidentally make changes to it, which would bypass the data saving code.  This
would also corrupt the setting when other parts of the PDK try to access the
same setting.

This commit duplicates the setting and default values which now isolates the
value from any changes.  And adds tests for the isolation.