(maint) Only call bundle when needed for validators
Previously PDK would call bundle, even if the validator doesn't need it, for
example internal validators like JSON or YAML.  This commit changes the CLI for
`pdk validate` to first check all of the requested validators to see if they
require a valid bundle and then only call bundle if needed.  This significantly
speeds up validation for validators like JSON or YAML syntax.