(PDK-1113) Add config helper method when value is nil
Previously the user needed to write verbose amounts of code to "do something if
and only if the setting value is not nil".  This commit adds a helper called
`with_scoped_value` which will yield the setting value if and only if the value
is not nil.  This commit then updates the PDK to use this helper and adds tests
for this new helper method.