(PDK-1614) Add project.environment settings
Now that the PDK can read the environment.conf file, it can be updated to
display that the user, and be consumed by the Control Repo context. This commit:

* Adds a helper method to PDK::ControlRepo to create an an object to read/edit
  an arbitrary environment.conf file.  The helper method also adds the default
  settings as per Puppet documentation
* Updates the PDK::Config class to take in a PDK Context. The configuration
  settings then are mutated depending on the Context, for example, the
  'puppet.environment.*' settings now exist if the user is in a Control Repo,
  and won't if they're in a Module.

Note that because PDK::Config now depends on a PDK::Context object being
passed in (or default via PDK.context), any method calls to configuration
information (ala PDK.config) are not allowed and can cause circular
references and stack overflows. Thus there is a helper class to create the
environment.conf IniFile object.

Note that right now there is no project-level PDK configuration file therefore
the mock configuration RSpec context does not mock this and that the
PDK::Config.project_config method has no filename associated with it.  Future
commits may add this feature.