(PDK-1618) Add controlrepo feature flag
The new Control Repo support is being put behind a feature flag so that users
can opt in, while the PDK team complete the feature.

This commit changes the context to detection to only detect Control Repos if the
flag is set.  The default behaviour of assuming everything is a module is still
maintained, unless the flag is set.