(maint) Validators should use unique target list
Previously the parse_targets method could return targets more than once which
causes the Puppet Parser Validator to fail due to duplicate definitions.  This
commit updates the parse_targets method to ensure that the targets are unique.

Ideally this should be done by only adding to the target list if it didn't
already exist, however due to the globbing, it's much easier to just use the
.uniq method on the array.