(#821) Don't use regex to match JSON
Regex + infinitely recursive data structures like JSON == sadness. It
can be done, but its just not the right tool for the job.

As the JSON formatter from rspec always outputs a condensed object, we
can say with surety that each document will be on its own line. So
rather than trying to scan for JSON objects using a regex, we can just
split the output by lines and look for lines that start with `{` and
end with `}` and attempt to parse them.