(PDK-1635) Add refactored Template Fetcher, Template Dir and Renderer
Previously the templating engine code was strewn about in multiple locations in
the PDK.  This commit aggregates and then abstracts the logic into the
PDK::Template namespace.

* (No change) A PDK::Util::TemplateURI points to where a remote template is located

* A PDK::Template::Fetcher will download the template that the TemplateURI
  points to

* PDK::Template::TemplateDir is created which is mainly just a helper class to
  make rendering easier. The TemplateDir contains an instance of a
  PDK::Template::Renderer.

* A PDK::Template::Renderer will render files from the template.  The PDK can
  then use the rendered output to either prompt users about impending changes
  or just apply the changes immediately.

Summary: A URI points a template, which a Fetcher fetches the template from,
which a TemplateDir creates a Renderer from the fetched template, and then the
Renderer can render files in the template.

Both the Fetcher and Renderer use a factory pattern to create the most
appropriate object instance. (PDK::Template::Fetcher.instance and
PDK::Template::Renderer.instance respectively) For example;

A PDK::Template::Fetcher:Git is created when a TemplateURI points to a remote
git repository.

Note that the only renderer available, called V1, is a refactored version of the
existing PDK template rendering code.  Later commits may add updated or even
completely new renderers