(PDK-1635) Add update manager printer
Previously only the PDK Module update and convert methods had access to print
out a summary from the UpdateManager class. However the UpdateManager is a very
versatile class and should be used more throughout the PDK.  This commit adds
an UpdateManagerPrinter which centralises the logic to print out the summary
of an UpdateManager. The printer can also accept a :tense option which will
changed the wording from future tense (e.g. "Files to be added") to past tense
("e.g. Files were added"). This is useful in CLI so the same printer can be
used before and after changes are applied.

Using a centralised module also means a consistent user experience no matter
where in the PDK codebase the manager is used.

This commit also adds tests, modified from the pdk/module/convert test suite.