(GH-874) Use PDK Context root for PDK Convert and Update
Previously the pdk update and convert CLI commands had a strange failure mode
where if a "module" was nested inside another "module" but only the parent had
a metadata.json. When a user tried to update or convert the child module, the
files would be put in the parent.  This was due to the CLI using the lax
detection (in_module_root?) method to "find" the module, but use the strict
detection method (module_root) during the actual conversion.

This commit updates these CLI logic to use the new PDK.context object which
is far more consistent.  This commit also adds a test for these scenarios in
PDK update and convert.